Introduce build step for docs.js, enable code splitting and increased testability
