+++ .gitignore
@@ -19,6 +19,13 @@ metastore_db
 html_docs/
 vendor
 
+# Compiled docs.js
+.cache/
+resources/web/docs.js
+resources/web/docs.js.map
+resources/web/tests/
+
+
 # Comes from running rspec tests
 .rspec_status
