+++ package.json
@@ -1,11 +1,15 @@
 {
   "license": "SEE LICENSE IN README.asciidoc",
   "devDependencies": {
+    "dedent": "^0.7.0",
     "jest": "^24.7.1",
     "jquery": "^1.11.3",
-    "dedent": "^0.7.0"
+    "parcel": "^1.12.3"
   },
   "scripts": {
     "test": "jest"
+  },
+  "dependencies": {
+    "js-cookie": "^2.1.0"
   }
 }
