+++ resources/web/docs_js/__tests__/docs.test.js
@@ -1,7 +1,9 @@
-const dedent = require('dedent');
-window.jQuery = require('jquery');
-const docs = require('../docs.js');
-docs.init_strings('en');
+const jQuery = require('jquery');
+import dedent from "dedent";
+import * as docs from "../index";
+import {lang_strings} from "../utils";
+
+const LangStrings = lang_strings('en');
 
 function pageWithConsole(name, consoleText, extraTextAssertions) {
   describe(name, () => {
@@ -18,7 +20,7 @@ function pageWithConsole(name, consoleText, extraTextAssertions) {
         </div>
       `;
 
-      docs.init_console_widgets('localhost:5601/app/kibana#/dev_tools/console');
+      docs.init_console_widgets('localhost:5601/app/kibana#/dev_tools/console', LangStrings);
       copyAsCurl = jQuery('.copy_as_curl');
     });
 
@@ -38,10 +40,10 @@ function pageWithConsole(name, consoleText, extraTextAssertions) {
       });
       describe('the copied text', () => {
         test('starts with curl', () => {
-          expect(document.copied).toMatch(/^curl/)
+          expect(document.copied).toMatch(/^curl/);
         });
         test('includes a method', () => {
-          expect(document.copied).toMatch(/-X/)
+          expect(document.copied).toMatch(/-X/);
         });
         test('includes the Content-Type', () => {
           expect(document.copied).toMatch(/-H 'Content-Type: application\/json'/);
@@ -59,7 +61,7 @@ describe('console widget', () => {
   describe('Copy as cURL button', () => {
     pageWithConsole('a snippet without a body', 'GET /_cat/health?v', () => {
       test('includes the corrent method', () => {
-        expect(document.copied).toMatch(/-X GET/)
+        expect(document.copied).toMatch(/-X GET/);
       });
       test('includes the url', () => {
         expect(document.copied).toMatch(/"localhost:9200\/_cat\/health\?v"/);
@@ -76,7 +78,7 @@ describe('console widget', () => {
     `
     pageWithConsole('a snippet with a body', withBody, () => {
       test('includes the method', () => {
-        expect(document.copied).toMatch(/-X PUT/)
+        expect(document.copied).toMatch(/-X PUT/);
       });
       test('includes the url', () => {
         expect(document.copied).toMatch(/"localhost:9200\/twitter\/_doc\/1"/);
@@ -109,7 +111,7 @@ function describeInitHeaders(name, guideBody, onThisPageAssertions) {
       `;
 
       const rightCol = jQuery('#right_col');
-      docs.init_headers(rightCol);
+      docs.init_headers(rightCol, LangStrings);
     });
 
     describe('the "On This Page" section', () => {
@@ -159,7 +161,7 @@ describe('On This Page', () => {
   describeInitHeaders('for page with one subsection', oneSubsection, () => {
     existsAssertions();
     test('contains a link to the subsection header', () => {
-      const link = jQuery('#this_page a[href="#nrt"]')
+      const link = jQuery('#this_page a[href="#nrt"]');
       expect(link).toHaveLength(1);
       expect(link.text().trim()).toEqual('Near Realtime (NRT)');
     });
@@ -167,12 +169,12 @@ describe('On This Page', () => {
   describeInitHeaders('for page with two subsections', twoSubsections, () => {
     existsAssertions();
     test('contains a link to the first subsection header', () => {
-      const link = jQuery('#this_page a[href="#nrt"]')
+      const link = jQuery('#this_page a[href="#nrt"]');
       expect(link).toHaveLength(1);
       expect(link.text().trim()).toEqual('Near Realtime (NRT)');
     });
     test('contains a link to the second subsection header', () => {
-      const link = jQuery('#this_page a[href="#cluster"]')
+      const link = jQuery('#this_page a[href="#cluster"]');
       expect(link).toHaveLength(1);
       expect(link.text().trim()).toEqual('Cluster');
     });
